size_t to int64_t with asserts

+++ QuickSort.h
@@ -8,12 +8,12 @@ class QuickSort
 	void qSortNaive(std::vector<int>& a) { // facade 
 		qSortNaive(a, 0, a.size() - 1);
 	}
-	void qSortNaive(std::vector<int>& a, size_t beg, size_t end); // implementation
+	void qSortNaive(std::vector<int>& a, int64_t beg, int64_t end); // implementation
 
 	void qSortAVX(std::vector<int>& a) { // facade 
 		qSortAVX(a, 0, a.size());
 	}
-	void qSortAVX(std::vector<int>& a, size_t beg, size_t end); // implementation
+	void qSortAVX(std::vector<int>& a, int64_t beg, int64_t end); // implementation
 
 	double meassuredSort(void(QuickSort::* sortFunc)(std::vector<int>&), std::vector<int>& a);
 
@@ -30,5 +30,5 @@ public:
 	}
 
 	bool prove(std::vector<int>& a);
-	std::vector<int> createRandomData(uint64_t size);
+	std::vector<int> createRandomData(int64_t size);
 };
